+++ crates/objc2/src/rc/weak.rs
@@ -1,8 +1,9 @@
 use alloc::boxed::Box;
 use core::cell::UnsafeCell;
 use core::fmt;
-use core::marker::PhantomData;
-use core::ptr;
+use core::marker::{PhantomData, PhantomPinned};
+use core::pin::Pin;
+use core::ptr::{self, NonNull};
 use std::panic::{RefUnwindSafe, UnwindSafe};
 
 use super::Retained;
@@ -11,9 +12,8 @@ use crate::{ffi, Message};
 
 /// A weak pointer to an Objective-C reference counted object.
 ///
-/// The object is allowed to be deallocated while the weak pointer is alive,
-/// though the backing allocation for the object can only be released once all
-/// weak pointers are gone.
+/// Once the object is deallocated, all remaining weak pointers to it is set
+/// to `NULL`.
 ///
 /// Useful for breaking reference cycles and safely checking whether an
 /// object has been deallocated.
@@ -25,28 +25,47 @@ use crate::{ffi, Message};
 /// library, and hence is only usable on types where `Retained<T>` acts like
 /// [`sync::Arc`], a.k.a. on non-mutable types.
 ///
+///
+/// # Pinning
+///
+/// This is a [pinned][core::pin] object, meaning it can generally only be
+/// used in places where you know that it won't move. This can be accomplished
+/// using the wrappers `Pin<Box<Weak<T>>>` or `Pin<&Weak<T>>`.
+///
+/// In instance variables, you will need to wrap the ivar access with
+/// `Pin::new_unchecked`, which is safe because instance variables are pinned.
+///
+///
+/// # Memory layout
+///
+/// This is guaranteed to have the same size and alignment as a pointer to the
+/// object, `*const T`. In Objective-C terms, this has the same layout as a
+/// `__weak` variable (and is ABI-compatible with that).
+///
 /// [`sync::Weak`]: std::sync::Weak
 /// [`sync::Arc`]: std::sync::Arc
-#[repr(transparent)] // This is not a public guarantee
+#[repr(transparent)]
 #[doc(alias = "WeakId")] // Previous name
+#[cfg_attr(
+    feature = "unstable-coerce-pointee",
+    derive(std::marker::CoercePointee)
+)]
 pub struct Weak<T: ?Sized> {
-    /// We give the runtime the address to this box, so that it can modify it
-    /// even if the `Weak` is moved.
+    /// The runtime holds the address of this field, and will zero it when the
+    /// object is deallocated.
     ///
-    /// Loading may modify the pointer through a shared reference, so we use
-    /// an UnsafeCell to get a *mut without self being mutable.
+    /// Since this means the pointer may be modified through a shared
+    /// reference, we use an `UnsafeCell`.
     ///
-    /// Remember that any thread may actually modify the inner value
-    /// concurrently, but as long as we only use it through the `objc_XXXWeak`
-    /// methods, all access is behind a lock.
-    ///
-    /// TODO: Verify the need for UnsafeCell?
-    /// TODO: Investigate if we can avoid some allocations using `Pin`.
-    /// TODO: Add derive(CoercePointee) once this doesn't Box internally.
-    inner: Box<UnsafeCell<*mut AnyObject>>,
-    /// Weak inherits variance, dropck and various marker traits from
-    /// `Retained<T>`.
-    item: PhantomData<Retained<T>>,
+    /// Note that any thread may actually modify the inner value concurrently,
+    /// but as long as we only use it through the `objc_XXXWeak` methods, all
+    /// access is behind a lock.
+    inner: UnsafeCell<*mut AnyObject>,
+    /// The runtime holds a reference to this type, so it cannot be moved
+    /// after it has been constructed.
+    pinned: PhantomPinned,
+    /// Covariant over `T`, and necessary for dropck.
+    item: PhantomData<T>,
 }
 
 /// Fully-deprecated type-alias to [`Weak`].
@@ -54,62 +73,272 @@ pub struct Weak<T: ?Sized> {
 pub type WeakId<T> = Weak<T>;
 
 impl<T: Message> Weak<T> {
-    /// Construct a new weak pointer that references the given object.
-    #[doc(alias = "objc_initWeak")]
+    /// A pointer to the inner object pointer.
+    ///
+    /// The outer pointer is a valid `__weak` object, or points to `NULL`, for
+    /// as long as the [`Weak`] is alive.
+    ///
+    /// The pointer is not synchronized, so reading from it is only valid when
+    /// you know that nothing else is touching it.
     #[inline]
-    pub fn new(obj: &T) -> Self {
-        // SAFETY: Pointer is valid since it came from a reference.
-        unsafe { Self::new_inner(obj) }
+    #[allow(dead_code)] // Difficult semantics, I'd rather not expose this yet.
+    fn as_ptr(&self) -> NonNull<*mut T> {
+        NonNull::new(self.inner.get().cast()).unwrap()
     }
 
-    /// Construct a new weak pointer that references the given [`Retained`].
-    #[doc(alias = "objc_initWeak")]
-    #[deprecated = "use `Weak::from_retained` instead"]
+    /// Constructs a new weak pointer that doesn't reference any object.
+    ///
+    /// # Example
+    ///
+    /// Place a weak pointer in a `static`.
+    ///
+    /// ```
+    /// use std::pin::Pin;
+    /// use objc2::rc::Weak;
+    /// # // New class that is Send + Sync.
+    /// # objc2::define_class!(
+    /// #     #[derive(Debug, PartialEq, Eq, Hash)]
+    /// #     #[unsafe(super(objc2::runtime::NSObject))]
+    /// #     #[name = "WeakUUID"]
+    /// #     struct NSUUID;
+    /// # );
+    /// #
+    /// # impl NSUUID {
+    /// #     objc2::extern_methods!(
+    /// #         #[unsafe(method(new))]
+    /// #         fn UUID() -> objc2::rc::Retained<Self>;
+    /// #     );
+    /// # }
+    /// # #[cfg(requires_foundation)]
+    /// use objc2_foundation::NSUUID;
+    ///
+    /// static WEAK: Weak<NSUUID> = Weak::empty();
+    ///
+    /// // Initially nothing is stored in the weak pointer.
+    /// assert_eq!(WEAK.load(), None);
+    ///
+    /// // But we can set the weak pointer to an object.
+    /// let obj = NSUUID::UUID();
+    /// Pin::static_ref(&WEAK).store(Some(&obj));
+    ///
+    /// // Such that it now loads that object.
+    /// assert_eq!(WEAK.load().as_ref(), Some(&obj));
+    ///
+    /// // Until the object is no longer alive.
+    /// drop(obj);
+    /// assert_eq!(WEAK.load(), None);
+    /// ```
     #[inline]
-    pub fn from_id(obj: &Retained<T>) -> Self {
-        Self::from_retained(obj)
-    }
-
-    /// Construct a new weak pointer that references the given [`Retained`].
-    #[doc(alias = "objc_initWeak")]
-    #[inline]
-    pub fn from_retained(obj: &Retained<T>) -> Self {
-        // SAFETY: Pointer is valid since it came from `Retained`.
-        unsafe { Self::new_inner(Retained::as_ptr(obj)) }
-    }
-
-    /// Raw constructor.
-    ///
-    ///
-    /// # Safety
-    ///
-    /// The object must be valid or null.
-    unsafe fn new_inner(obj: *const T) -> Self {
-        let inner = Box::new(UnsafeCell::new(ptr::null_mut()));
-        // SAFETY: `ptr` will never move, and the caller verifies `obj`
-        let _ = unsafe { ffi::objc_initWeak(inner.get(), (obj as *mut T).cast()) };
+    pub const fn empty() -> Self {
+        // SAFETY: The pointer is null, which is valid for us to pass to all
+        // the `objc_*Weak` functions.
         Self {
-            inner,
+            inner: UnsafeCell::new(ptr::null_mut()),
+            pinned: PhantomPinned,
             item: PhantomData,
         }
     }
 
+    /// Construct a new boxed weak pointer.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use objc2::rc::Weak;
+    /// use objc2::runtime::NSObject;
+    ///
+    /// let obj = NSObject::new();
+    /// let weak = Weak::new(&*obj);
+    /// assert_eq!(weak.load(), Some(obj));
+    /// ```
+    #[doc(alias = "objc_initWeak")]
+    // We do two operations (allocate Box and init weak), but we probably want
+    // this to be inlined regardless, as that'd allow reordering and re-use of
+    // the allocation (though fairly unlikely).
+    #[inline]
+    pub fn new(obj: &T) -> Pin<Box<Self>> {
+        let obj = obj as *const T as *mut T;
+
+        let boxed = Box::pin(Self::default());
+        // SAFETY:
+        // - The weak pointer is newly initialized, and will never move
+        //   because it is pinned.
+        // - The object pointer is valid since it came from a reference.
+        let _ = unsafe { ffi::objc_initWeak(boxed.inner.get(), obj.cast()) };
+        boxed
+    }
+
+    /// Construct a new boxed weak pointer that references the given [`Retained`].
+    #[doc(alias = "objc_initWeak")]
+    #[deprecated = "use `Weak::from_retained` instead"]
+    #[inline]
+    pub fn from_id(obj: &Retained<T>) -> Pin<Box<Self>> {
+        Self::from_retained(obj)
+    }
+
+    /// Construct a new boxed weak pointer that references the given
+    /// [`Retained`] object.
+    ///
+    /// Convenience alias for [`Weak::new`].
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use objc2::rc::Weak;
+    /// use objc2::runtime::NSObject;
+    ///
+    /// let obj = NSObject::new();
+    /// let weak = Weak::from_retained(&obj); // No `Deref` needed
+    /// assert_eq!(weak.load(), Some(obj));
+    /// ```
+    #[doc(alias = "objc_initWeak")]
+    #[inline]
+    pub fn from_retained(obj: &Retained<T>) -> Pin<Box<Self>> {
+        Self::new(obj)
+    }
+
     /// Load the object into an [`Retained`] if it still exists.
     ///
-    /// Returns [`None`] if the object has been deallocated, or the `Weak`
-    /// was created with [`Default::default`].
+    /// Returns [`None`] if:
+    /// - The object has been deallocated.
+    /// - The weak pointer was created with [`Weak::empty()`] and not yet set.
+    /// - The weak pointer was cleared with [`weak.store(None)`][Self::store].
+    ///
+    /// # Example
+    ///
+    /// Load an object from a weak pointer as long as it's still alive.
+    ///
+    /// ```
+    /// use objc2::rc::Weak;
+    /// use objc2::runtime::NSObject;
+    ///
+    /// let obj = NSObject::new();
+    /// let weak = Weak::from_retained(&obj);
+    ///
+    /// // Loading an object that is alive returns `Some`.
+    /// assert_eq!(weak.load().as_ref(), Some(&obj));
+    ///
+    /// drop(obj);
+    /// // The object is no longer alive, so now loading returns `None`.
+    /// assert_eq!(weak.load(), None);
+    /// ```
+    ///
+    /// Load an zero-initialized weak pointer.
+    ///
+    /// ```
+    /// use objc2::rc::Weak;
+    /// use objc2::runtime::NSObject;
+    ///
+    /// let obj = NSObject::new();
+    /// let weak = Weak::<NSObject>::empty();
+    /// assert_eq!(weak.load(), None);
+    /// ```
     #[doc(alias = "retain")]
     #[doc(alias = "objc_loadWeak")]
     #[doc(alias = "objc_loadWeakRetained")]
     #[inline]
     pub fn load(&self) -> Option<Retained<T>> {
-        let ptr = self.inner.get();
-        let obj = unsafe { ffi::objc_loadWeakRetained(ptr) }.cast();
-        // SAFETY: The object has +1 retain count
+        // SAFETY: The weak pointer is either NULL (newly initialized) or
+        // contains a weak pointer.
+        //
+        // NOTE: We don't need `self: Pin<&Self>`, since all functions that
+        // store something into the pointer require pinning. Thus the only
+        // time where `self` would be un-pinned would be if it was
+        // zero-initialized.
+        //
+        // This makes usage slightly nicer (we don't need `.as_ref()`), and
+        let obj = unsafe { ffi::objc_loadWeakRetained(self.inner.get()) }.cast();
+        // SAFETY: The object has +1 retain count from ^.
         unsafe { Retained::from_raw(obj) }
     }
 
     // TODO: Add `autorelease(&self, pool) -> Option<&T>` using `objc_loadWeak`?
+
+    /// Make a boxed copy of the weak pointer that points to the same object.
+    #[doc(alias = "objc_copyWeak")]
+    #[inline]
+    pub fn copy(self: Pin<&Self>) -> Pin<Box<Self>> {
+        let boxed = Box::pin(Self::empty());
+        // SAFETY:
+        // - The source pointer is either NULL (newly initialized) or contains
+        //   a weak pointer, and it will never move because it is pinned.
+        // - The destination pointer is newly initialized, and will never move
+        //   because it is pinned.
+        unsafe { ffi::objc_copyWeak(boxed.inner.get(), self.inner.get()) };
+        boxed
+    }
+
+    /// Set the weak pointer to a new object.
+    ///
+    /// You can explicitly set the value to [`None`], though it is rarely that
+    /// useful, as that will will be done automatically once the object is
+    /// deallocated.
+    ///
+    ///
+    /// # Examples
+    ///
+    /// Store a value into an empty weak pointer.
+    ///
+    /// ```
+    /// use std::pin::pin;
+    /// use objc2::rc::Weak;
+    /// use objc2::runtime::NSObject;
+    ///
+    /// let weak = pin!(Weak::default());
+    /// assert_eq!(weak.load(), None);
+    ///
+    /// let obj = NSObject::new();
+    /// weak.as_ref().store(Some(&*obj));
+    /// assert_eq!(weak.load(), Some(obj));
+    /// ```
+    ///
+    /// Overwrite a boxed weak pointer.
+    ///
+    /// ```
+    /// use objc2::rc::Weak;
+    /// use objc2::runtime::NSObject;
+    ///
+    /// let obj1 = NSObject::new();
+    /// let weak = Weak::from_retained(&obj1);
+    ///
+    /// let obj2 = NSObject::new();
+    /// weak.as_ref().store(Some(&obj2));
+    ///
+    /// assert_eq!(weak.load(), Some(obj2));
+    /// ```
+    ///
+    /// Explicitly set the weak pointer to [`None`].
+    ///
+    /// ```
+    /// use objc2::rc::Weak;
+    /// use objc2::runtime::NSObject;
+    ///
+    /// let obj = NSObject::new();
+    /// let weak = Weak::from_retained(&obj);
+    ///
+    /// weak.as_ref().store(None);
+    /// assert_eq!(weak.load(), None);
+    /// ```
+    #[doc(alias = "objc_storeWeak")]
+    #[inline]
+    pub fn store(self: Pin<&Self>, obj: Option<&T>) {
+        let obj = obj
+            .map(|obj| obj as *const T as *mut T)
+            .unwrap_or_else(ptr::null_mut);
+
+        // NOTE: We don't use `objc_destroyWeak` on `NULL` pointers, as that
+        // is not guaranteed to be thread-safe.
+        // If we wanted to do that instead, we'd need `Pin<&mut Self>`.
+
+        // SAFETY:
+        // - The weak pointer is either NULL (newly initialized) or contains a
+        //   weak pointer, and it will never move because it is pinned.
+        // - The object pointer is either NULL or valid since it came from a
+        //   reference.
+        let _ = unsafe { ffi::objc_storeWeak(self.inner.get(), obj.cast()) };
+    }
+
+    // TODO: objc_moveWeak? Probably needs better pin ergonomics.
 }
 
 impl<T: ?Sized> Drop for Weak<T> {
@@ -117,21 +346,20 @@ impl<T: ?Sized> Drop for Weak<T> {
     #[doc(alias = "objc_destroyWeak")]
     #[inline]
     fn drop(&mut self) {
+        // SAFETY: The weak pointer is either valid or contains NULL.
+        //
+        // We also know that it was never moved since its construction because
+        // `Self` is `!Unpin` (which means we can safely pass).
         unsafe { ffi::objc_destroyWeak(self.inner.get()) }
     }
 }
 
 // TODO: Add ?Sized
-impl<T: Message> Clone for Weak<T> {
-    /// Make a clone of the weak pointer that points to the same object.
+impl<T: Message> Clone for Pin<Box<Weak<T>>> {
     #[doc(alias = "objc_copyWeak")]
+    #[inline]
     fn clone(&self) -> Self {
-        let ptr = Box::new(UnsafeCell::new(ptr::null_mut()));
-        unsafe { ffi::objc_copyWeak(ptr.get(), self.inner.get()) };
-        Self {
-            inner: ptr,
-            item: PhantomData,
-        }
+        self.as_ref().copy()
     }
 }
 
@@ -139,11 +367,10 @@ impl<T: Message> Clone for Weak<T> {
 impl<T: Message> Default for Weak<T> {
     /// Constructs a new weak pointer that doesn't reference any object.
     ///
-    /// Calling [`Self::load`] on the return value always gives [`None`].
+    /// This is equivalent to [`Weak::empty()`].
     #[inline]
     fn default() -> Self {
-        // SAFETY: The pointer is null
-        unsafe { Self::new_inner(ptr::null()) }
+        Self::empty()
     }
 }
 
@@ -157,13 +384,14 @@ impl<T: ?Sized> fmt::Debug for Weak<T> {
 }
 
 // SAFETY: Same as `std::sync::Weak<T>`.
+// The `objc_*Weak` methods use a lock internally, so it's safe to load this
+// from multiple threads, even while the underlying object may be deallocated.
 unsafe impl<T: ?Sized + Sync + Send> Sync for Weak<T> {}
 
 // SAFETY: Same as `std::sync::Weak<T>`.
+// The `objc_*Weak` methods use a lock internally, so it's safe to load this
+// from multiple threads, even while the underlying object may be deallocated.
 unsafe impl<T: ?Sized + Sync + Send> Send for Weak<T> {}
-
-// Same as `std::sync::Weak<T>`.
-impl<T: ?Sized> Unpin for Weak<T> {}
 
 // Same as `std::sync::Weak<T>`.
 impl<T: ?Sized + RefUnwindSafe> RefUnwindSafe for Weak<T> {}
@@ -171,21 +399,21 @@ impl<T: ?Sized + RefUnwindSafe> RefUnwindSafe for Weak<T> {}
 // Same as `std::sync::Weak<T>`.
 impl<T: ?Sized + RefUnwindSafe> UnwindSafe for Weak<T> {}
 
-impl<T: Message> From<&T> for Weak<T> {
+impl<T: Message> From<&T> for Pin<Box<Weak<T>>> {
     #[inline]
     fn from(obj: &T) -> Self {
         Weak::new(obj)
     }
 }
 
-impl<T: Message> From<&Retained<T>> for Weak<T> {
+impl<T: Message> From<&Retained<T>> for Pin<Box<Weak<T>>> {
     #[inline]
     fn from(obj: &Retained<T>) -> Self {
         Weak::from_retained(obj)
     }
 }
 
-impl<T: Message> From<Retained<T>> for Weak<T> {
+impl<T: Message> From<Retained<T>> for Pin<Box<Weak<T>>> {
     #[inline]
     fn from(obj: Retained<T>) -> Self {
         Weak::from_retained(&obj)
@@ -199,14 +427,18 @@ mod tests {
     use super::*;
     use crate::rc::{RcTestObject, ThreadTestData};
     use crate::runtime::NSObject;
+    use crate::{define_class, msg_send, AnyThread, Ivars};
 
     #[test]
     fn test_weak() {
         let obj = RcTestObject::new();
         let mut expected = ThreadTestData::current();
 
-        let weak = Weak::from(&obj);
+        let weak = Weak::from_retained(&obj);
         expected.assert_current();
+
+        // A weak object was registered at the pointer.
+        assert!(!unsafe { weak.inner.get().read() }.is_null());
 
         let strong = weak.load().unwrap();
         expected.try_retain += 1;
@@ -218,6 +450,9 @@ mod tests {
         expected.release += 2;
         expected.drop += 1;
         expected.assert_current();
+
+        // The object is deallocated, which sets weak pointers to null.
+        assert!(unsafe { weak.inner.get().read() }.is_null());
 
         if cfg!(not(feature = "gnustep-1-7")) {
             // This loads the object on GNUStep for some reason??
@@ -234,7 +469,7 @@ mod tests {
         let obj = RcTestObject::new();
         let mut expected = ThreadTestData::current();
 
-        let weak = Weak::from(&obj);
+        let weak = Weak::from_retained(&obj);
         expected.assert_current();
 
         let weak2 = weak.clone();
@@ -261,7 +496,12 @@ mod tests {
 
     #[test]
     fn test_weak_default() {
-        let weak: Weak<RcTestObject> = Weak::default();
+        let weak: Weak<RcTestObject> = Default::default();
+        assert!(weak.load().is_none());
+        drop(weak);
+
+        let weak: Pin<Box<Weak<RcTestObject>>> = Default::default();
+        assert!(weak.clone().load().is_none());
         assert!(weak.load().is_none());
         drop(weak);
     }
@@ -280,9 +520,51 @@ mod tests {
 
     #[test]
     fn test_size_of() {
-        assert_eq!(
-            mem::size_of::<Option<Weak<NSObject>>>(),
-            mem::size_of::<*const ()>()
-        );
+        let ptr_size = mem::size_of::<*const ()>();
+        assert_eq!(mem::size_of::<Weak<NSObject>>(), ptr_size);
+        assert_ne!(mem::size_of::<Option<Weak<NSObject>>>(), ptr_size);
+    }
+
+    static_assertions::assert_not_impl_any!(Weak<NSObject>: Unpin);
+
+    define_class!(
+        #[unsafe(super(NSObject))]
+        struct WeakClass {
+            boxed: Pin<Box<Weak<RcTestObject>>>,
+            plain: Weak<RcTestObject>,
+        }
+    );
+
+    impl WeakClass {
+        fn new(obj: &RcTestObject) -> Retained<Self> {
+            let this = Self::alloc().set_ivars(Ivars::<Self> {
+                boxed: Weak::new(obj),
+                plain: Weak::empty(),
+            });
+            // Call `NSObject`'s `init` method.
+            let this: Retained<Self> = unsafe { msg_send![super(this), init] };
+            // SAFETY: Ivars are pinned.
+            unsafe { Pin::new_unchecked(this.plain()) }.store(Some(obj));
+            this
+        }
+    }
+
+    #[test]
+    fn defined() {
+        let obj = RcTestObject::new();
+        let mut expected = ThreadTestData::current();
+
+        let x = WeakClass::new(&obj);
+        expected.assert_current();
+
+        assert_eq!(x.boxed().load().as_ref(), Some(&obj));
+        expected.try_retain += 1;
+        expected.release += 1;
+        expected.assert_current();
+
+        assert_eq!(x.plain().load().as_ref(), Some(&obj));
+        expected.try_retain += 1;
+        expected.release += 1;
+        expected.assert_current();
     }
 }